added session to conftest, removed make_domain_bib
